Convert `src/redux/uniswap` to TypeScript and update related file.
